Change fact handling into a setting
This changes the behavior of fact handling. This makes it easier to
change without calling configure_beaker. Doing so plays better with
modulesync.

It also moves it into its own file. This keeps the global namespace
cleaner.
