+++ examples/defaults_facts/manifests/init.pp
@@ -1,0 +1,6 @@
+class defaults_facts {
+  file { "/voxpupuli-acceptance-test":
+    ensure  => 'file',
+    content => "Current test: ${module_name}\nThe answer: ${facts['the_answer']}\n",
+  }
+}
