+++ examples/defaults_facts/metadata.json
@@ -1,0 +1,14 @@
+{
+  "name": "voxpupuli_acceptance_tests-defaults_facts",
+  "version": "0.0.1",
+  "author": "Vox Pupuli",
+  "license": "Apache-2.0",
+  "summary": "The voxpupuli-acceptance test suite defaults_facts",
+  "description": "DESCRIBE THE TEST SUITE HERE",
+  "source": "https://github.com/voxpupuli/voxpupuli-acceptance",
+  "project_page": "https://github.com/voxpupuli/voxpupuli-acceptance",
+  "issues_url": "https://github.com/voxpupuli/voxpupuli-acceptance/issues",
+  "dependencies": [],
+  "requirements": [],
+  "operatingsystem_support": []
+}
