+++ examples/defaults_facts/spec/spec_helper_acceptance.rb
@@ -1,0 +1,5 @@
+require 'voxpupuli/acceptance/spec_helper_acceptance'
+
+ENV['BEAKER_FACTER_THE_ANSWER'] = '42'
+
+configure_beaker
