+++ lib/voxpupuli/acceptance/spec_helper_acceptance.rb
@@ -1,33 +1,6 @@
 require_relative 'examples'
 
-ENV_VAR_PREFIX = 'BEAKER_FACTER_'
-FACT_FILE = '/etc/facter/facts.d/voxpupuli-acceptance-env.json'
-
-def beaker_facts_from_env
-  facts = {}
-
-  ENV.each do |var, value|
-    next unless var.start_with?(ENV_VAR_PREFIX)
-
-    fact = var.sub(ENV_VAR_PREFIX, '').downcase
-    facts[fact] = value
-  end
-
-  facts
-end
-
-def write_beaker_facts_on(hosts)
-  beaker_facts = beaker_facts_from_env
-
-  if beaker_facts.any?
-    require 'json'
-    on(hosts, "mkdir -p #{File.dirname(FACT_FILE)} && cat <<VOXPUPULI_BEAKER_ENV_VARS > #{FACT_FILE}\n#{beaker_facts.to_json}\nVOXPUPULI_BEAKER_ENV_VARS")
-  else
-    on(hosts, "rm -f #{FACT_FILE}")
-  end
-end
-
-def configure_beaker(modules: :metadata, configure_facts_from_env: true, &block)
+def configure_beaker(modules: :metadata, &block)
   ENV['PUPPET_INSTALL_TYPE'] ||= 'agent'
   ENV['BEAKER_PUPPET_COLLECTION'] ||= 'puppet6'
   ENV['BEAKER_debug'] ||= 'true'
@@ -65,7 +38,10 @@ def configure_beaker(modules: :metadata, configure_facts_from_env: true, &block)
         Voxpupuli::Acceptance::Fixtures.install_fixture_modules_on(hosts, fixture_modules)
       end
 
-      write_beaker_facts_on(hosts) if configure_facts_from_env
+      if RSpec.configuration.suite_configure_facts_from_env
+        require_relative 'facts'
+        Voxpupuli::Acceptance::Facts.write_beaker_facts_on(hosts)
+      end
 
       if RSpec.configuration.suite_hiera?
         hiera_data_dir = RSpec.configuration.suite_hiera_data_dir
@@ -92,6 +68,9 @@ def configure_beaker(modules: :metadata, configure_facts_from_env: true, &block)
 end
 
 RSpec.configure do |c|
+  # Fact handling
+  c.add_setting :suite_configure_facts_from_env, default: true
+
   # Hiera settings
   c.add_setting :suite_hiera, default: true
   c.add_setting :suite_hiera_data_dir, default: File.join('spec', 'acceptance', 'hieradata')
